+++ dave/common/logger.py
@@ -2,6 +2,8 @@ import logging
 import os
 from .singleton import SingletonMeta
 
+from . import server_type as st
+
 
 class Logger(
     metaclass=SingletonMeta,
@@ -11,7 +13,32 @@ class Logger(
         self.__logger = logging.getLogger("dave")
         loglevel = os.environ.get("DAVE_LOGLEVEL", "INFO").upper()
         self.__logger.setLevel(loglevel)
-        # self.__logger.config(level=loglevel, format="%(levelname)s: %(message)s")
 
-    def get(self) -> logging.Logger:
-        return self.__logger
+    def info(self, message: str):
+        if st.SERVER_TYPE == st.ServerType.GDB:
+            import gdb
+            gdb.write(f"INFO: {message}")
+        else:
+            self.__logger.info(message)
+
+    def warning(self, message: str):
+        if st.SERVER_TYPE == st.ServerType.GDB:
+            import gdb
+            gdb.write(f"WARNING: {message}")
+        else:
+            self.__logger.warning(message)
+
+    def debug(self, message: str):
+        if st.SERVER_TYPE == st.ServerType.GDB and self.__logger.level == "debug":
+            import gdb
+            gdb.write(f"DEBUG: {message}")
+        else:
+            self.__logger.debug(message)
+
+    def error(self, message: str):
+        if st.SERVER_TYPE == st.ServerType.GDB:
+            import gdb
+            gdb.write(f"ERROR: {message}")
+        else:
+            self.__logger.error(message)
+