+++ dave/server/container_factory.py
@@ -60,13 +60,13 @@ class ContainerFactory(metaclass=SingletonMeta):
     def build(
         self, dbg_value: Any, typename: str, varname: str, dims: List[int] = []
     ) -> Container:
-        Logger().get().debug(f"Building {varname} from type |{typename}|")
+        Logger().debug(f"Building {varname} from type |{typename}|")
 
         # First we check if it is a 1D container
         try:
             return self.build_1D(dbg_value, typename, varname, dims)
         except (ContainerError, TypeError):
-            Logger().get().debug(f"{typename} is not a valid 1D container")
+            Logger().debug(f"{typename} is not a valid 1D container")
             pass
 
         # Then we check for 2D containers