+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support for custom containers classes
 ## Changed
 - Improved gdb frame change detection
+- Improced gdb logging
 ## Fixed
 - `freeze`, `concat`, `delete` command printing their argument on lldb
 - `freeze`, `concat` commands not affecting the gui corresponding switches